vertical_2_collateral-verification/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
motor==3.4.0
pymongo==4.7.2
pyjwt==2.8.0
bcrypt==4.0.1
python-multipart==0.0.9
pydantic==2.7.1
python-dotenv==1.0.1
reportlab==4.2.0